+++ internal/asm/arm64/golang_asm.go
@@ -7,7 +7,7 @@ import (
 	"github.com/twitchyliquid64/golang-asm/obj"
 	"github.com/twitchyliquid64/golang-asm/obj/arm64"
 
-	"github.com/tetratelabs/wazero/internal/wasm/jit/asm"
+	"github.com/tetratelabs/wazero/internal/asm"
 )
 
 // assemblerGoAsmImpl implements Assembler for golang-asm library.
