+++ internal/asm/golang_asm.go
