+++ internal/asm/impl.go
