+++ internal/wasm/jit/jit_arm64_impl.go
@@ -14,8 +14,8 @@ import (
 	"math"
 	"unsafe"
 
+	"github.com/tetratelabs/wazero/internal/asm"
 	wasm "github.com/tetratelabs/wazero/internal/wasm"
-	"github.com/tetratelabs/wazero/internal/wasm/jit/asm"
 	"github.com/tetratelabs/wazero/internal/wasm/jit/asm/arm64"
 	"github.com/tetratelabs/wazero/internal/wazeroir"
 )
